Relocate skills + tools to upstream Tier C layout
Upstream standardized on a Tier C bundle layout while this branch
was off in the wilderness:

  scilink/skills/<category>/<name>/<name>.md   (skill markdown)
  scilink/skills/<category>/<name>/<name>.py   (per-skill helper)
  scilink/skills/<category>/<name>/__init__.py
  scilink/skills/_shared/<helper>.py            (cross-skill helpers)
  scilink/tools/                                (mostly gone -- tools
                                                 now live with their skill,
                                                 cf. force_field/amber/amber.py)

Move:
  scilink/skills/molecular_dynamics/lammps.md
    -> scilink/skills/molecular_dynamics/lammps/lammps.md
  scilink/skills/machine_learning_potentials/{general,mace}.md
    -> scilink/skills/machine_learning_potentials/{general,mace}/<name>.md
  scilink/tools/molecular_dynamics/lammps_tools.py
    -> scilink/skills/molecular_dynamics/lammps/lammps.py
  scilink/tools/mlip_tools.py
    -> scilink/skills/_shared/mlip_tools.py
    (multi-backend dispatcher used by all MLIP skill bundles, so it
     belongs in _shared not under any one skill)

Add per-bundle __init__.py files so the loader's path resolution
finds the bundles.

Update import sites:
  md_simulation_agent.py:
    from ...skills.molecular_dynamics.lammps import lammps as lammps_tools
  mlip_agent.py:
    from ...skills._shared import mlip_tools
  tests/test_lammps_skill/test_lammps_tools.py: same import update

Expose the new agents in sim_agents/__init__.py:
  SimulationAgent (base), MDSimulationAgent, MLIPAgent.
LAMMPSSimulationAgent stays as the backward-compat alias around
MDSimulationAgent.

This is a pure layout move; no functional changes. 119 tests pass
under tests/test_lammps_skill/ + tests/test_skill_loader.py.

The architectural intent (generic MD agent + per-package skills and
tools) is now realized in code: the agent has no LAMMPS-specific
methods; LAMMPS-specific behavior comes from the skill markdown +
the LAMMPS tool module discovered via TOOL_REGISTRY. Adding GROMACS
or OpenMM is now "drop in a skill bundle + a tools module + register
it" rather than "subclass LAMMPSSimulationAgent."
