+++ scilink/agents/sim_agents/__init__.py
@@ -6,6 +6,7 @@ from .vasp_quality import VaspQualityAgent
 from .base_agent import SimulationAgent
 from .md_simulation_agent import MDSimulationAgent
 from .lammps_agent import LAMMPSSimulationAgent
+from .mlip_agent import MLIPAgent
 from .lammps_updater import LAMMPSUpdater
 from .force_field_agent import ForceFieldAgent
 from .lammps_analysis import LAMMPSAnalysisAgent
